Update implementation to accomodate hacking around youtubes SPA
The youtube page is only 'loaded' once, every subsequent request afterwards uses their SPA templating (probably angular) and swaps out / replaces DOM nodes. Once a youtube video is loaded, that 'page' exists in the background and is reused, just check for changes on that node, while if it hasnt been loaded yet, observe the body until a node that resembles it exists.